Add RTC and time support to the kernel
This commit updates the Makefile to include rtc.o and time.o. It introduces the struct timeval in defs.h, adds the gettimeofday function, and integrates RTC memory mapping in vm.c. The main function is modified to demonstrate the use of gettimeofday, printing the current time. Additionally, a new time_t type is defined in types.h.