Add time manipulation functions and struct tm to the kernel
This commit introduces the struct tm for time representation and implements the mktime and localtime_r functions for converting between time representations. The main function is updated to demonstrate the usage of these new time functions, printing the formatted local time and the corresponding time_t value. Additionally, the gettimeofday function is utilized to retrieve the current time.